Day 16 part 2

+++ src/main/java/com/jacoblucas/adventofcode2021/day16/Packet.java
@@ -21,7 +21,7 @@ public abstract class Packet {
 
     public abstract int getVersionSum();
 
-    public abstract int get();
+    public abstract long get();
 
     @Override
     public boolean equals(Object o) {
